Indexing MacOS computer now works
Modified the MacLocalIndexer script to be inline with the new functionality of IndalekoMachineConfig and large changes made to Indaleko Class. The MacMachineConfig revamped to be inline with rest of Indaleko scripts. Most of the __init__ script for activity/collectors is commented out as I can not seem to download the winsdk python package which is required to run the __init___. Since this init is quintessential to run the super class script Indaleko.py, I commented out the lines 36-41 - which is all about activity monitoring.